Move encrypt and decrypt parameter types into cryptography model header file.
