+++ sdk/keyvault/azure-security-keyvault-keys/src/private/remote_cryptography_client.hpp
@@ -17,7 +17,6 @@
 
 #include "azure/keyvault/keys/cryptography/cryptography_client_models.hpp"
 #include "azure/keyvault/keys/cryptography/cryptography_client_options.hpp"
-#include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
 #include "azure/keyvault/keys/key_client_models.hpp"
 
 #include <memory>
